feat: parallax homepage
