+++ components/page.module.css
@@ -1,0 +1,15 @@
+.page {
+  width: 100vw;
+  height: 100vh;
+  margin: 0;
+  padding: 0;
+  position: absolute;
+  top: 0;
+  left: 0;
+
+  color: var(--color-fg);
+}
+
+.sep {
+  margin-top: 4rem;
+}
