+++ next.config.js
@@ -2,6 +2,9 @@
 const nextConfig = {
   reactStrictMode: true,
   swcMinify: true,
-}
+  images: {
+    domains: ["media.discordapp.net"],
+  },
+};
 
-module.exports = nextConfig
+module.exports = nextConfig;
