+++ package.json
@@ -14,6 +14,7 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-is": "^18.2.0",
+    "react-scroll-parallax": "^3.3.1",
     "styled-components": "^5.3.6",
     "styled-icons": "^10.46.0",
     "victormono": "^1.5.4"
