+++ pages/_app.js
@@ -1,8 +1,13 @@
 import "victormono";
 import "../styles/globals.css";
+import Page from "../components/page.js";
 
 function MyApp({ Component, pageProps }) {
-  return <Component {...pageProps} />;
+  return (
+    <Page>
+      <Component {...pageProps} />
+    </Page>
+  );
 }
 
 export default MyApp;
