+++ styles/utils.module.css
@@ -1,0 +1,5 @@
+.bottom {
+  width: 100vw;
+  height: 1000px;
+  background-color: var(--color-bg);
+}
